+++ internal/broadcast/client.go
@@ -2,7 +2,9 @@ package broadcast
 
 import (
 	"encoding/json"
+	"fmt"
 	"log"
+	"reflect"
 
 	"github.com/gorilla/websocket"
 
@@ -51,6 +53,18 @@ func (c *Client) readMessages(cm *ConnectionManager) {
 			log.Printf("error unmarshaling json: %v", err)
 			continue
 		}
+
+		parsedPayload, err := parsePayload(event)
+		if err != nil {
+			fmt.Println(err)
+			continue
+		}
+
+		event = Event{
+			Type:    event.Type,
+			Payload: parsedPayload,
+		}
+
 		if err := cm.routeEvent(event, c); err != nil {
 			log.Printf("error routing event: %v", err)
 		}
@@ -71,25 +85,60 @@ func (c *Client) writeMessages(cm *ConnectionManager) {
 	}()
 
 	for {
-		select {
-		case message, ok := <-c.egress:
-			// Ok will be false Incase the egress channel is closed
-			if !ok {
-				// Manager has closed this connection channel, so communicate that to frontend
-				if err := c.connection.WriteMessage(websocket.CloseMessage, nil); err != nil {
-					// Log that the connection is closed and the reason
-					log.Println("connection closed: ", err)
-				}
-				// Return to close the goroutine
-				return
+
+		message, ok := <-c.egress
+		// Ok will be false Incase the egress channel is closed
+		if !ok {
+			// Manager has closed this connection channel, so communicate that to frontend
+			if err := c.connection.WriteMessage(websocket.CloseMessage, nil); err != nil {
+				// Log that the connection is closed and the reason
+				log.Println("connection closed: ", err)
 			}
-			// Write a Regular text message to the connection
-			data, _ := json.Marshal(message)
-			if err := c.connection.WriteMessage(websocket.TextMessage, data); err != nil {
-				log.Println(err)
-			}
-			log.Println("sent message")
+			// Return to close the goroutine
+			return
 		}
+		// Write a Regular text message to the connection
+		data, _ := json.Marshal(message)
+		if err := c.connection.WriteMessage(websocket.TextMessage, data); err != nil {
+			log.Println(err)
+		}
+		log.Println("sent message")
 
 	}
+}
+
+func parsePayload(event Event) (interface{}, error) {
+	switch event.Type {
+	case EventSendMessage:
+		payload, ok := event.Payload.(string)
+		if !ok {
+			return nil, fmt.Errorf("parsePaload: %v is of type %v not string", event.Payload, reflect.TypeOf(event.Payload))
+		}
+		return payload, nil
+	case EventReqMatch:
+		payload, ok := event.Payload.([]interface{})
+		if !ok {
+			return nil, fmt.Errorf("parsePaload: %v is of type %v not []interface{}", event.Payload, reflect.TypeOf(event.Payload))
+		}
+		payloadString := make([]string, len(payload))
+		for i, v := range payload {
+			str, ok := v.(string)
+			if !ok {
+				return nil, fmt.Errorf("type assertion failed for element: %v", v)
+			}
+			payloadString[i] = str
+		}
+		return payloadString, nil
+	case EventReqSessionEnd:
+		payload, ok := event.Payload.(struct{})
+		if !ok {
+			return nil, fmt.Errorf("parsePaload: %v is of type %v not struct", event.Payload, reflect.TypeOf(event.Payload))
+		}
+		return payload, nil
+
+	default:
+		fmt.Println("Unknown event type")
+	}
+	return nil, fmt.Errorf("parsePaload: %v is an invalid event type", event.Type)
+
 }