+++ internal/broadcast/handlers.go
@@ -1,11 +1,15 @@
 package broadcast
 
-import "fmt"
+import (
+	"fmt"
+
+	"github.com/majorbruteforce/hi-five/internal/matchmaker"
+)
 
 type EventHandler func(event Event, c *Client) error
 
 // setupEventHandlers configures and adds all handlers
-func (cm *ConnectionManager) setupEventHandlers() {
+func (cm *ConnectionManager) setupEventHandlers(m *matchmaker.MatchManager) {
 	cm.handlers[EventSendMessage] = func(e Event, c *Client) error {
 		if _, ok := cm.sessions[c]; ok {
 			msg := e.Payload
@@ -20,9 +24,7 @@ func (cm *ConnectionManager) setupEventHandlers() {
 	}
 
 	cm.handlers[EventReqMatch] = func(e Event, c *Client) error {
-		// m := matchmaker.NewManager()
-		// m.Ingress <- Candidate{ID: c.Id, Keywords: e.Payload}
-
+		m.Ingress <- matchmaker.Candidate{ID: c.ID, Keywords: e.Payload.([]string)}
 		return nil
 	}