+++ internal/matchmaker/manager.go
@@ -1,7 +1,7 @@
 package matchmaker
 
 import (
-	"container/list"
+	"fmt"
 	"time"
 )
 
@@ -12,39 +12,98 @@ type Candidate struct {
 
 const (
 	// Initiate batch matching after config.targetLength
-	StrategyBatchSize int = 0
+	StrategySize int = 0
 	// Initiate batch matching after config.targetTime
-	StrategyBatchTime int = 1
+	StrategyTime int = 1
 	// Initiate a continous match and dispatch strategy
 	StrategyRapid int = 2
 )
 
 type Config struct {
-	targetLength int
-	targetTime   time.Duration
-	strategy     int
+	TargetBufferSize int
+	TargetWaitTime   time.Duration
+	Strategy         int
 }
 
 type MatchManager struct {
-	queue   []*list.List
+	buffer  []Candidate
 	config  Config
 	Ingress chan Candidate
 	Egress  chan [][2]string
+	stop    chan struct{}
 }
 
 func NewManager(config Config) *MatchManager {
-	return &MatchManager{
-		queue:  make([]*list.List, 0),
-		config: config,
+	m := &MatchManager{
+		buffer:  make([]Candidate, 0),
+		config:  config,
+		Ingress: make(chan Candidate),
+		Egress:  make(chan [][2]string),
+		stop:    make(chan struct{}),
+	}
+	m.NewMatchMaker()
+	return m
+}
+
+func (m *MatchManager) NewMatchMaker() {
+
+	switch m.config.Strategy {
+	case StrategySize:
+		go m.runStrategySize()
+	case StrategyTime:
+		go m.runStrategyTime()
+	case StrategyRapid:
+		go m.runStrategyRapid()
+	default:
+		fmt.Println("Unknown strategy")
+	}
+	fmt.Println("New matchmaker running")
+}
+
+func (m *MatchManager) runStrategySize() {
+	for {
+		select {
+		case arrival, ok := <-m.Ingress:
+			if !ok {
+				fmt.Println("matchmaker ingress channel is closed and drained")
+				return
+			}
+			m.buffer = append(m.buffer, arrival)
+			if len(m.buffer) >= m.config.TargetBufferSize {
+				matches := MatchCandidatesBatch(m.buffer)
+				m.Egress <- matches
+				m.buffer = m.buffer[:0]
+			}
+		case <-m.stop:
+			return
+		}
 	}
 }
 
-func (m *MatchManager) NewMatchMaker() error {
-	// Accepts incoming Candidates through Ingress, creates matches
-	// and facilitates session creation.
-	//
-	// When target for the strategy is met, it sends the
-	// the list to MatchCandidatesBatch and pipes the result
-	// to broadcast.CreateBatchSessions
-	return nil
+func (m *MatchManager) runStrategyTime() {
+	ticker := time.NewTicker(m.config.TargetWaitTime)
+	defer ticker.Stop()
+
+	for {
+		select {
+		case arrival, ok := <-m.Ingress:
+			if !ok {
+				fmt.Println("matchmaker ingress channel is closed and drained")
+				return
+			}
+			m.buffer = append(m.buffer, arrival)
+		case <-ticker.C:
+			if len(m.buffer) > 0 {
+				matches := MatchCandidatesBatch(m.buffer)
+				m.Egress <- matches
+				m.buffer = m.buffer[:0]
+			}
+		case <-m.stop:
+			return
+		}
+	}
+}
+
+func (m *MatchManager) runStrategyRapid() {
+	// Implement
 }